update, delete data

+++ app/Http/Controllers/TodoController.php
@@ -16,18 +16,17 @@ class TodoController extends Controller
             ->get();
             // dd($todos);
 
-        return view('todo.index', compact('todos'));
+            $todosCompleted = Todo::where('user_id', Auth::user()->id)
+            ->where('is_done', true)
+            ->count();
+
+        return view('todo.index', compact('todos', 'todosCompleted'));
     }
 
 
     public function create()
     {
         return view('todo.create');
-    }
-
-    public function edit()
-    {
-        return view('todo.edit');
     }
 
     public function store(Request $request)
@@ -43,10 +42,88 @@ class TodoController extends Controller
         return redirect()->route('todo.index')->with('success', 'todo created successfully.');
     }
 
+
+
+    public function edit(Todo $todo)
+    {
+        if (Auth::user()->id == $todo->user_id) {
+            return view('todo.edit', compact('todo'));
+        } else {
+            return redirect()->route('todo.index')->with('danger', 'You are not authorized to edit this todo!');
+        }
+    }
+
+    public function update(Request $request, Todo $todo)
+    {
+        if ($todo->user_id != Auth::user()->id) {
+            return redirect()->route('todo.index')->with('danger', 'You are not authorized to update this todo!');
+        }
+
+        $request->validate([
+            'title' => 'required|max:255',
+        ]);
+
+        $todo->update([
+            'title' => ucfirst($request->title),
+        ]);
+
+        return redirect()->route('todo.index')->with('success', 'Todo updated successfully!');
+    }
+
+
+
+
+
+
+    public function complete(Todo $todo)
+    {
+        if ($todo->user_id != Auth::user()->id) {
+            abort(403);
+        }
+
+        $todo->is_done = true;
+        $todo->save();
+
+        return redirect()->route('todo.index')->with('success', 'Todo completed successfully!');
+    }
+
+    public function uncomplete(Todo $todo)
+    {
+        if ($todo->user_id != Auth::user()->id) {
+            abort(403);
+        }
+
+        $todo->is_done = false;
+        $todo->save();
+
+        return redirect()->route('todo.index')->with('success', 'Todo uncompleted successfully!');
+    }
+
+
+
+
+
+
     public function destroy(Todo $todo)
     {
-        $todo->delete();
+        if (Auth::user()->id == $todo->user_id) {
+            $todo->delete();
+            return redirect()->route('todo.index')->with('success', 'Todo deleted successfully!');
+        } else {
+            return redirect()->route('todo.index')->with('danger', 'You are not authorized to delete this todo!');
+        }
+    }
 
-        return redirect()->route('todo.index')->with('success', 'todo deleted successfully.');
+    public function destroyCompleted()
+    {
+        $todosCompleted = Todo::where('user_id', Auth::user()->id)
+            ->where('is_done', true)
+            ->get();
+
+        foreach ($todosCompleted as $todo) {
+            $todo->delete();
+        }
+
+        return redirect()->route('todo.index')->with('success', 'All completed todos deleted successfully!');
     }
 }
